hillel - move from helper to util
